requirements.txt
flask
gunicorn
pytest
flake8
requests
mysql-connector-python